+++ src/module/applications/ux/enrichers/lookup.mjs
@@ -26,13 +26,16 @@ export async function enricher(match, options) {
   const parsedConfig = parseConfig(config);
   parsedConfig._input = match[0];
 
+  parsedConfig.formulaPath = parsedConfig.path ?? parsedConfig.formula;
+
   for (const val of parsedConfig.values) {
-    if (["capitalize", "lowercase", "uppercase"].includes(val)) parsedConfig.style ??= val;
-    else if (val.startsWith("@")) parsedConfig.path ??= val;
-    else if (val === "formula") parsedConfig.formula = true;
+    const normalizedValue = val.toLowerCase();
+    if (["capitalize", "lowercase", "uppercase"].includes(normalizedValue)) parsedConfig.style ??= normalizedValue;
+    else if (normalizedValue === "evaluate") parsedConfig.evaluate = true;
+    else parsedConfig.formulaPath ??= val;
   }
 
-  if (!parsedConfig.path) {
+  if (!parsedConfig.formulaPath) {
     console.warn(`Lookup path must be defined to enrich ${config._input}.`);
     return null;
   }
@@ -41,9 +44,9 @@ export async function enricher(match, options) {
 
   /** @type {string | number} */
   let value;
-  if (parsedConfig.formula) value = ds.utils.evaluateFormula(parsedConfig.path, data, { contextName: "lookup" });
+  if (parsedConfig.evaluate) value = ds.utils.evaluateFormula(parsedConfig.formulaPath, data, { contextName: "lookup" });
   else {
-    value = foundry.utils.getProperty(data, parsedConfig.path.substring(1)) ?? fallback;
+    value = foundry.utils.getProperty(data, parsedConfig.formulaPath.substring(1)) ?? fallback;
 
     switch (parsedConfig.style) {
       case "capitalize":
@@ -62,7 +65,7 @@ export async function enricher(match, options) {
   span.classList.add("lookup-value");
   if (!value && (options.documents === false)) return null;
   else if (!value) span.classList.add("not-found");
-  span.innerText = value ?? parsedConfig.path;
+  span.innerText = value ?? parsedConfig.formulaPath;
   return span;
 }